+++ mobile/packages/footprint_flutter/CHANGELOG.md
@@ -9,3 +9,4 @@
 ## 1.0.2
 
 - Remove theme variable from appearance
+- Fix a bug that prevented the users to initialize the flow with auth token